Modified abs_get_pkgbuild() to be able to download from the Arch official svntogit repos. Runs as normal by default, but defaults to git if USE_GIT=1 is set
The abs tree is only updated/synced once a day, so this prevents a yaourt -Sb from downloading an outdated PKGBUILD and compiling it anyway.
E.g. ffmpeg 2.2.3-1 is available, but try to build from source a few hours after it is released, and yaourt would get the 2.2.2 PKGBUILD as abs is still out of date.

Signed-off-by: stevenhoneyman <stevenhoneyman@users.noreply.github.com>