Prefer $VISUAL over $EDITOR, prefer env vars to config.
Add a VISUAL option to yaourtrc and as an environment variable.
Prefer VISUAL to EDITOR always because it is meant to be more full
featured. EDITOR is still supported if VISUAL is unset.

Change the help messages to prefer VISUAL.

Also, prefer environment variables to yaourtrc always.
Document this change.