add a --confirm option to pacdiffviewer
Yaourt already has this flag, so pacdiffviewer should have it, too.
Especially, as the default answer to the question what to do with a file
is to take no action, which essentially turns pacdiffviewer into no-as
long as NOCONFIRM is set in yaourtrc A flag re-enabling the question
makes it usable again without unconditionally ignoring the
configuration.