Rename INSTALL to INSTALL_FLAG
Autoconf tries to get a BSD install command and use INSTALL variable to
store the result. So if INSTALL is set to something by yaourt, it will
make autoconf set install command pointing to this and end up in an
error during the package.
In order to avoid collision between -i option of major -P (pkgbuild)
with this INSTALL variable of autoconf, rename -i flag to INSTALL_FLAG.